+++ components/rendering/src/lib.rs
@@ -32,7 +32,7 @@ use errors::Result;
 pub use context::RenderContext;
 use markdown::markdown_to_html;
 pub use shortcode::render_shortcodes;
-pub use table_of_contents::Header;
+pub use table_of_contents::Heading;
 
 pub fn render_content(content: &str, context: &RenderContext) -> Result<markdown::Rendered> {
     // Don't do shortcodes if there is nothing like a shortcode in the content