+++ components/templates/src/global_fns/mod.rs
@@ -33,8 +33,12 @@ impl TeraFn for Trans {
         let key = required_arg!(String, args.get("key"), "`trans` requires a `key` argument.");
         let lang = optional_arg!(String, args.get("lang"), "`trans`: `lang` must be a string.")
             .unwrap_or_else(|| self.config.default_language.clone());
-        let translations = &self.config.translations[lang.as_str()];
-        Ok(to_value(&translations[key.as_str()]).unwrap())
+
+        let term = self.config.get_translation(lang, key).map_err(|e| {
+            Error::chain("Failed to retreive term translation", e)
+        })?;
+
+        Ok(to_value(term).unwrap())
     }
 }
 
@@ -94,8 +98,8 @@ impl ResizeImage {
     }
 }
 
-static DEFAULT_OP: &'static str = "fill";
-static DEFAULT_FMT: &'static str = "auto";
+static DEFAULT_OP: &str = "fill";
+static DEFAULT_FMT: &str = "auto";
 const DEFAULT_Q: u8 = 75;
 
 impl TeraFn for ResizeImage {
@@ -203,8 +207,9 @@ impl TeraFn for GetTaxonomyUrl {
             args.get("name"),
             "`get_taxonomy_url` requires a `name` argument with a string value"
         );
-        let lang = optional_arg!(String, args.get("lang"), "`get_taxonomy`: `lang` must be a string")
-            .unwrap_or_else(|| self.default_lang.clone());
+        let lang =
+            optional_arg!(String, args.get("lang"), "`get_taxonomy`: `lang` must be a string")
+                .unwrap_or_else(|| self.default_lang.clone());
 
         let container = match self.taxonomies.get(&format!("{}-{}", kind, lang)) {
             Some(c) => c,
@@ -296,7 +301,11 @@ pub struct GetTaxonomy {
     default_lang: String,
 }
 impl GetTaxonomy {
-    pub fn new(default_lang: &str, all_taxonomies: Vec<Taxonomy>, library: Arc<RwLock<Library>>) -> Self {
+    pub fn new(
+        default_lang: &str,
+        all_taxonomies: Vec<Taxonomy>,
+        library: Arc<RwLock<Library>>,
+    ) -> Self {
         let mut taxonomies = HashMap::new();
         for taxo in all_taxonomies {
             taxonomies.insert(format!("{}-{}", taxo.kind.name, taxo.kind.lang), taxo);
@@ -312,8 +321,9 @@ impl TeraFn for GetTaxonomy {
             "`get_taxonomy` requires a `kind` argument with a string value"
         );
 
-        let lang = optional_arg!(String, args.get("lang"), "`get_taxonomy`: `lang` must be a string")
-            .unwrap_or_else(|| self.default_lang.clone());
+        let lang =
+            optional_arg!(String, args.get("lang"), "`get_taxonomy`: `lang` must be a string")
+                .unwrap_or_else(|| self.default_lang.clone());
 
         match self.taxonomies.get(&format!("{}-{}", kind, lang)) {
             Some(t) => Ok(to_value(t.to_serialized(&self.library.read().unwrap())).unwrap()),
@@ -409,8 +419,8 @@ mod tests {
         let tags_fr = Taxonomy { kind: taxo_config_fr, items: vec![tag_fr] };
 
         let taxonomies = vec![tags.clone(), tags_fr.clone()];
-        let static_fn = GetTaxonomy::new(&config.default_language, taxonomies.clone(), library.clone())
-            ;
+        let static_fn =
+            GetTaxonomy::new(&config.default_language, taxonomies.clone(), library.clone());
         // can find it correctly
         let mut args = HashMap::new();
         args.insert("kind".to_string(), to_value("tags").unwrap());
@@ -501,9 +511,8 @@ mod tests {
         assert!(static_fn.call(&args).is_err());
     }
 
-    #[test]
-    fn can_translate_a_string() {
-        let trans_config = r#"
+
+    const TRANS_CONFIG: &str = r#"
 base_url = "https://remplace-par-ton-url.fr"
 default_language = "fr"
 
@@ -513,10 +522,11 @@ title = "Un titre"
 
 [translations.en]
 title = "A title"
-
         "#;
 
-        let config = Config::parse(trans_config).unwrap();
+    #[test]
+    fn can_translate_a_string() {
+        let config = Config::parse(TRANS_CONFIG).unwrap();
         let static_fn = Trans::new(config);
         let mut args = HashMap::new();
 
@@ -528,5 +538,27 @@ title = "A title"
 
         args.insert("lang".to_string(), to_value("fr").unwrap());
         assert_eq!(static_fn.call(&args).unwrap(), "Un titre");
+    }
+
+    #[test]
+    fn error_on_absent_translation_lang() {
+        let mut args = HashMap::new();
+        args.insert("lang".to_string(), to_value("absent").unwrap());
+        args.insert("key".to_string(), to_value("title").unwrap());
+
+        let config = Config::parse(TRANS_CONFIG).unwrap();
+        let error = Trans::new(config).call(&args).unwrap_err();
+        assert_eq!("Failed to retreive term translation", format!("{}", error));
+    }
+
+    #[test]
+    fn error_on_absent_translation_key() {
+        let mut args = HashMap::new();
+        args.insert("lang".to_string(), to_value("en").unwrap());
+        args.insert("key".to_string(), to_value("absent").unwrap());
+
+        let config = Config::parse(TRANS_CONFIG).unwrap();
+        let error = Trans::new(config).call(&args).unwrap_err();
+        assert_eq!("Failed to retreive term translation", format!("{}", error));
     }
 }