+++ src/cmd/serve.rs
@@ -52,6 +52,7 @@ use rebuild;
 enum ChangeKind {
     Content,
     Templates,
+    Themes,
     StaticFiles,
     Sass,
     Config,
@@ -117,6 +118,7 @@ fn create_new_site(
     output_dir: &str,
     base_url: &str,
     config_file: &str,
+    include_drafts: bool,
 ) -> Result<(Site, String)> {
     let mut site = Site::new(env::current_dir().unwrap(), config_file)?;
 
@@ -131,6 +133,9 @@ fn create_new_site(
     site.config.enable_serve_mode();
     site.set_base_url(base_url);
     site.set_output_path(output_dir);
+    if include_drafts {
+        site.include_drafts();
+    }
     site.load()?;
     site.enable_live_reload(port);
     console::notify_site_size(&site);
@@ -147,14 +152,17 @@ pub fn serve(
     config_file: &str,
     watch_only: bool,
     open: bool,
+    include_drafts: bool,
 ) -> Result<()> {
     let start = Instant::now();
-    let (mut site, address) = create_new_site(interface, port, output_dir, base_url, config_file)?;
+    let (mut site, address) =
+        create_new_site(interface, port, output_dir, base_url, config_file, include_drafts)?;
     console::report_elapsed_time(start);
 
     // Setup watchers
     let mut watching_static = false;
     let mut watching_templates = false;
+    let mut watching_themes = false;
     let (tx, rx) = channel();
     let mut watcher = watcher(tx, Duration::from_secs(1)).unwrap();
     watcher
@@ -176,6 +184,13 @@ pub fn serve(
         watcher
             .watch("templates/", RecursiveMode::Recursive)
             .map_err(|e| ZolaError::chain("Can't watch the `templates` folder.", e))?;
+    }
+
+    if Path::new("themes").exists() {
+        watching_themes = true;
+        watcher
+            .watch("themes/", RecursiveMode::Recursive)
+            .map_err(|e| ZolaError::chain("Can't watch the `themes` folder.", e))?;
     }
 
     // Sass support is optional so don't make it an error to no have a sass folder
@@ -247,6 +262,9 @@ pub fn serve(
     }
     if watching_templates {
         watchers.push("templates");
+    }
+    if watching_themes {
+        watchers.push("themes");
     }
     if site.config.compile_sass {
         watchers.push("sass");
@@ -362,6 +380,22 @@ pub fn serve(
                             ChangeKind::Templates => reload_templates(&mut site, &path),
                             ChangeKind::StaticFiles => copy_static(&site, &path, &partial_path),
                             ChangeKind::Sass => reload_sass(&site, &path, &partial_path),
+                            ChangeKind::Themes => {
+                                console::info(
+                                    "-> Themes changed. The whole site will be reloaded.",
+                                );
+                                site = create_new_site(
+                                    interface,
+                                    port,
+                                    output_dir,
+                                    base_url,
+                                    config_file,
+                                    include_drafts,
+                                )
+                                .unwrap()
+                                .0;
+                                rebuild_done_handling(&broadcaster, Ok(()), "/x.js");
+                            }
                             ChangeKind::Config => {
                                 console::info("-> Config changed. The whole site will be reloaded. The browser needs to be refreshed to make the changes visible.");
                                 site = create_new_site(
@@ -370,6 +404,7 @@ pub fn serve(
                                     output_dir,
                                     base_url,
                                     config_file,
+                                    include_drafts,
                                 )
                                 .unwrap()
                                 .0;
@@ -406,6 +441,22 @@ pub fn serve(
                             (ChangeKind::Templates, _) => reload_templates(&mut site, &path),
                             (ChangeKind::StaticFiles, p) => copy_static(&site, &path, &p),
                             (ChangeKind::Sass, p) => reload_sass(&site, &path, &p),
+                            (ChangeKind::Themes, _) => {
+                                console::info(
+                                    "-> Themes changed. The whole site will be reloaded.",
+                                );
+                                site = create_new_site(
+                                    interface,
+                                    port,
+                                    output_dir,
+                                    base_url,
+                                    config_file,
+                                    include_drafts,
+                                )
+                                .unwrap()
+                                .0;
+                                rebuild_done_handling(&broadcaster, Ok(()), "/x.js");
+                            }
                             (ChangeKind::Config, _) => {
                                 console::info("-> Config changed. The whole site will be reloaded. The browser needs to be refreshed to make the changes visible.");
                                 site = create_new_site(
@@ -414,6 +465,7 @@ pub fn serve(
                                     output_dir,
                                     base_url,
                                     config_file,
+                                    include_drafts,
                                 )
                                 .unwrap()
                                 .0;
@@ -471,6 +523,8 @@ fn detect_change_kind(pwd: &Path, path: &Path) -> (ChangeKind, PathBuf) {
 
     let change_kind = if partial_path.starts_with("/templates") {
         ChangeKind::Templates
+    } else if partial_path.starts_with("/themes") {
+        ChangeKind::Themes
     } else if partial_path.starts_with("/content") {
         ChangeKind::Content
     } else if partial_path.starts_with("/static") {
@@ -526,6 +580,11 @@ mod tests {
                 (ChangeKind::Templates, PathBuf::from("/templates/hello.html")),
                 Path::new("/home/vincent/site"),
                 Path::new("/home/vincent/site/templates/hello.html"),
+            ),
+            (
+                (ChangeKind::Themes, PathBuf::from("/themes/hello.html")),
+                Path::new("/home/vincent/site"),
+                Path::new("/home/vincent/site/themes/hello.html"),
             ),
             (
                 (ChangeKind::StaticFiles, PathBuf::from("/static/site.css")),